+++ src/components/AdminPane/Manage/TaskPropertyStyleRules/TaskPropertyStyleRules.js
@@ -3,7 +3,6 @@ import { FormattedMessage } from 'react-intl'
 import _map from 'lodash/map'
 import _head from 'lodash/head'
 import _isEmpty from 'lodash/isEmpty'
-import MarkdownContent from '../../../MarkdownContent/MarkdownContent'
 import SvgSymbol from '../../../SvgSymbol/SvgSymbol'
 import TaskPropertyQueryBuilder
        from '../../../TaskPropertyQueryBuilder/TaskPropertyQueryBuilder'
@@ -135,12 +134,18 @@ export default class TaskPropertyStyleRules extends Component {
     })
 
     return (
-      <div className="">
+      <div>
         <div className="mr-font-medium mr-text-yellow mr-uppercase mr-mb-2 mr-text-xl">
           <FormattedMessage {...messages.stylesHeader} />
         </div>
-        <div className="mr-mb-6">
-          <MarkdownContent markdown={this.props.intl.formatMessage(messages.stylesTip)} />
+        <div className="mr-mb-6 mr-links-green-light">
+          <a
+            href={`${process.env.REACT_APP_DOCS_URL}/documentation/styling-task-features/`}
+            target="_blank"
+            rel="noopener noreferrer"
+          >
+            <FormattedMessage {...messages.stylesDoc} />
+          </a>
         </div>
 
         <div className="mr-my-4 mr-text-white">