+++ src/components/HOCs/WithCommandInterpreter/WithCommandInterpreter.js
@@ -7,6 +7,9 @@ import _omit from 'lodash/omit'
 import _find from 'lodash/find'
 import _get from 'lodash/get'
 import _debounce from 'lodash/debounce'
+import _trim from 'lodash/trim'
+import _toNumber from 'lodash/toNumber'
+import _isNaN from 'lodash/isNaN'
 import { fetchPlaceLocation } from '../../../services/Place/Place'
 import WithErrors from '../WithErrors/WithErrors'
 import AppErrors from '../../../services/Error/AppErrors'
@@ -20,11 +23,12 @@ import { SEARCH_TYPE_PROJECT } from '../../SearchTypeFilter/SearchTypeFilter'
  *    m/  => Execute a map bounds search with either a bounding box or a centerpoint
  *    n/  => Execute a nominatim search and move map bounds
  *    p/  => Execute project name search
+ *    i/  => Execute a search by challenge id (exact matches only)
  *    s/ or default => Execute a standard search query
  *
  * @author [Kelli Rotstan](https://github.com/krotstan)
  */
-const WithCommandInterpreter = function(WrappedComponent) {
+const WithCommandInterpreter = function(WrappedComponent, acceptedCommands = null) {
   return class extends Component {
     state = {
       commandString: null,
@@ -35,9 +39,10 @@ const WithCommandInterpreter = function(WrappedComponent) {
     executeSearch = (commandString, searchType) => {
       // executeCommmand either runs the command or runs a standard challenge
       // search. It returns true if it ran a search, false if it ran a command
-      const wasStandardSearch = executeCommand(this.props, commandString, searchType, false, false)
+      const wasStandardSearch = executeCommand(this.props, commandString, searchType,
+                                               false, false, acceptedCommands)
       this.setState({
-        commandString: wasStandardSearch ? null : commandString,
+        commandString: commandString,
         searchType: searchType,
         searchActive: wasStandardSearch,
       })
@@ -46,9 +51,7 @@ const WithCommandInterpreter = function(WrappedComponent) {
     clearSearch = () => {
       // Temporary: until we add an Advanced Search dialog where a user can
       // clear a project search filter, we need to do it explicitly here
-      this.props.removeSearchFilters(['query'])
-      this.props.removeSearchFilters(['project'])
-      this.props.removeSearchFilters(['searchType'])
+      this.props.removeSearchFilters(['query', 'project', 'searchType', 'challengeId'])
 
       this.props.clearSearch()
       this.setState({commandString: null, searchType: null, searchActive: true})
@@ -56,7 +59,8 @@ const WithCommandInterpreter = function(WrappedComponent) {
 
     deactivate = () => {
       executeCommand(this.props, this.state.commandString, this.state.searchType,
-                     (loading) => this.setState({mapLoading: loading}), true)
+                     (loading) => this.setState({mapLoading: loading}), true,
+                     acceptedCommands)
     }
 
     componentDidUpdate(prevProps) {
@@ -93,6 +97,19 @@ const WithCommandInterpreter = function(WrappedComponent) {
 }
 
 /**
+ * Tests if the short code is supported and will add an error if it is not.
+ **/
+const isCommandSupported = (code, acceptedCommands, props) => {
+  if (acceptedCommands && acceptedCommands.indexOf(code) === -1) {
+    props.addErrorWithDetails(AppErrors.search.notSupported, code + "/")
+    return false
+  }
+  else {
+    return true
+  }
+}
+
+/**
  * Executes the appropriate search type based on the start of the query string.
  * If isComplete is set to true then the user has indicated the search string is
  * complete; otherwise user may be in the middle of typing/constructing their
@@ -100,7 +117,8 @@ const WithCommandInterpreter = function(WrappedComponent) {
  *
  * @return boolean - Whether this was a typical search or a command search
  */
-export const executeCommand = (props, commandString, searchType, setLoading, isComplete=false) => {
+export const executeCommand = (props, commandString, searchType, setLoading,
+                               isComplete=false, acceptedCommands) => {
   if (searchType === SEARCH_TYPE_PROJECT) {
     props.setSearchFilters({project: commandString})
     return false
@@ -109,29 +127,49 @@ export const executeCommand = (props, commandString, searchType, setLoading, isC
   const command = commandString && commandString.length >= 2 ? commandString.substring(0, 2) : null
   let query = commandString ? commandString.substring(2) : commandString
 
-  // Temporary: until we add an Advanced Search dialog where a user can clear a
-  // project search filter, we need to do it explicitly here if needed
-  if (command !== '/p') {
-    props.removeSearchFilters(['project'])
-  }
-
   switch(command) {
     case 'm/':
-      props.setSearch("")  // We need to clear the initial 'm' from the query
-      if (isComplete && query.length > 0) {
-        debouncedMapSearch(props, query, setLoading)
+      props.setSearch("") // We need to clear the initial 'm' from the query
+      if (isCommandSupported('m', acceptedCommands, props)) {
+        if (isComplete && query.length > 0) {
+          debouncedMapSearch(props, query, setLoading)
+        }
+      }
+      else {
+        props.setSearchFilters({})
       }
       return false
     case 'n/':
       props.setSearch("") // We need to clear the initial 'n' from the query
-      if (isComplete && query.length > 0) {
-        debouncedPlaceSearch(props, query, setLoading)
+      if (isCommandSupported('n', acceptedCommands, props)) {
+        if (isComplete && query.length > 0) {
+          debouncedPlaceSearch(props, query, setLoading)
+        }
+      }
+      else {
+        props.setSearchFilters({})
       }
       return false
     case 'p/':
       props.setSearch("") // We need to clear the initial 'p' from the query
-      if (query.length > 0) {
-        props.setSearchFilters({project: query})
+      if (isCommandSupported('p', acceptedCommands, props)) {
+        if (query.length > 0) {
+          props.setSearchFilters({project: query})
+        }
+      }
+      else {
+        props.setSearchFilters({})
+      }
+      return false
+    case 'i/':
+      props.setSearch("") // We need to clear the initial 'i' from the query
+      if (isCommandSupported('i', acceptedCommands, props)) {
+        if (query.length > 0 && !_isNaN(_toNumber(_trim(query)))) {
+          props.setSearchFilters({challengeId: _trim(query)})
+        }
+      }
+      else {
+        props.setSearchFilters({})
       }
       return false
     case 's/':
@@ -139,6 +177,8 @@ export const executeCommand = (props, commandString, searchType, setLoading, isC
       if (command !== 's/') {
         query = commandString
       }
+      // Remove any lingering search filters.
+      props.removeSearchFilters(['project', 'challengeId'])
 
       // Standard search query
       props.setSearch(query)
@@ -247,5 +287,5 @@ WithCommandInterpreter.propTypes = {
   clearSearch: PropTypes.func.isRequired,
 }
 
-export default WrappedComponent =>
-  WithErrors(WithCommandInterpreter(WrappedComponent))
+export default (WrappedComponent, acceptedCommands) =>
+  WithErrors(WithCommandInterpreter(WrappedComponent, acceptedCommands))