+++ src/components/HOCs/WithOSMUserSearch/WithOSMUserSearch.js
@@ -3,7 +3,7 @@ import _isEmpty from 'lodash/isEmpty'
 import _filter from 'lodash/filter'
 import _startsWith from 'lodash/startsWith'
 import _debounce from 'lodash/debounce'
-import { findUser } from '../../../services/User/User'
+import { findUser, findPreferredUsers } from '../../../services/User/User'
 
 /**
  * WithOSMUserSearch provides a findUser function to the wrapped component that allows
@@ -21,23 +21,36 @@ const WithOSMUserSearch = function(WrappedComponent) {
     /**
      * @private
      */
-    performSearch = _debounce(username => {
+    performSearch = _debounce((username, taskId) => {
       this.setState({isSearchingOSMUsers: true})
 
-      findUser(username).then(results => {
-        this.setState({isSearchingOSMUsers: false, osmUserResults: results})
-      })
+      if (_isEmpty(username)) {
+        findPreferredUsers(username, taskId).then(preferredResults => {
+          this.setState({isSearchingOSMUsers: false,
+                         osmUserResults: [],
+                         osmUserPreferredResults: preferredResults})
+        })
+      }
+      else {
+        findUser(username).then(results => {
+          if (taskId) {
+            findPreferredUsers(username, taskId).then(preferredResults => {
+              this.setState({isSearchingOSMUsers: false,
+                             osmUserResults: results,
+                             osmUserPreferredResults: preferredResults})
+            })
+          }
+          else {
+            this.setState({isSearchingOSMUsers: false, osmUserResults: results})
+          }
+        })
+      }
     }, 1000, {leading: true})
 
     /**
      * Initiates search for user with the given username or username fragment.
      */
     searchOSMUser = username => {
-      if (_isEmpty(username)) {
-        this.setState({osmUserResults: []})
-        return
-      }
-
       // Start off by filtering our existing search results so that we don't continue
       // to show results that no longer match the new username.
       this.setState({
@@ -46,7 +59,7 @@ const WithOSMUserSearch = function(WrappedComponent) {
           _startsWith(result.displayName.toLowerCase(), username.toLowerCase()))
       })
 
-      this.performSearch(username)
+      this.performSearch(username, this.props.taskId)
     }
 
     osmUserKey = osmUser => osmUser.osmId
@@ -57,6 +70,7 @@ const WithOSMUserSearch = function(WrappedComponent) {
       return (
         <WrappedComponent isSearching={this.state.isSearchingOSMUsers}
                           searchResults={this.state.osmUserResults}
+                          preferredResults={this.state.osmUserPreferredResults}
                           search={this.searchOSMUser}
                           resultKey={this.osmUserKey}
                           resultLabel={this.osmUserLabel}