+++ src/components/TaskConfirmationModal/TaskConfirmationModal.js
@@ -60,7 +60,9 @@ export class TaskConfirmationModal extends Component {
   }
 
   componentDidMount(prevProps, prevState) {
-    this.commentInputRef.current.focus()
+    if (this.commentInputRef.current) {
+      this.commentInputRef.current.focus()
+    }
 
     this.props.pauseKeyboardShortcuts()
     this.props.activateKeyboardShortcut(
@@ -69,8 +71,9 @@ export class TaskConfirmationModal extends Component {
       this.handleKeyboardShortcuts
     )
 
-    if (this.props.needsResponses && _isEmpty(this.props.completionResponses)) {
-      this.setState({showInstructions: true})
+    if (this.props.needsResponses && _isEmpty(this.props.completionResponses) &&
+        this.props.status !== TaskStatus.skipped) {
+      this.setState({showInstructions: true, instructionsContinue: true})
     }
   }
 
@@ -215,6 +218,7 @@ export class TaskConfirmationModal extends Component {
                         placeholder={applyingTagChanges ? '' : this.props.intl.formatMessage(messages.placeholder)}
                         value={this.props.comment}
                         commentChanged={this.props.setComment}
+                        taskId={this.props.task.id}
                       />
                     </div>
                     <KeywordAutosuggestInput
@@ -302,7 +306,7 @@ export class TaskConfirmationModal extends Component {
                         </label>
                       </div>
                       <div className="mr-text-green-lighter mr-text-center mr-mt-4 hover:mr-text-white mr-cursor-pointer mr-text-xs">
-                        <div onClick={() => this.setState({showInstructions: true})}>
+                        <div onClick={() => this.setState({showInstructions: true, instructionsContinue: false})}>
                           <FormattedMessage {...messages.viewInstructions} />
                         </div>
                       </div>
@@ -442,7 +446,9 @@ export class TaskConfirmationModal extends Component {
           {!this.props.inReview && this.state.showInstructions && _isUndefined(this.props.needsRevised) &&
             <InstructionsOverlay
               {...this.props}
-              close={() => this.setState({showInstructions: false})}
+              close={() => this.setState({showInstructions: false, instructionsContinue: false})}
+              closeMessage={this.state.instructionsContinue ?
+                messages.instructionsContinueLabel : messages.closeInstructionsLabel}
             />
           }
         </Modal>