+++ src/pages/Review/TasksReview/Messages.js
@@ -129,6 +129,11 @@ export default defineMessages({
     defaultMessage: "Reviewer",
   },
 
+  additionalReviewersLabel: {
+    id: "Review.fields.additionalReviewers.label",
+    defaultMessage: "Additional Reviewers",
+  },
+
   mappedOnLabel: {
     id: "Review.fields.mappedOn.label",
     defaultMessage: "Mapped On",
@@ -147,6 +152,11 @@ export default defineMessages({
   reviewAgainTaskLabel: {
     id: "Review.TaskAnalysisTable.controls.reviewAgainTask.label",
     defaultMessage: "Review Revision",
+  },
+
+  metaReviewTaskLabel: {
+    id: "Review.TaskAnalysisTable.controls.metaReviewTask.label",
+    defaultMessage: "Review further",
   },
 
   resolveTaskLabel: {