+++ src/pages/Review/TasksReview/TasksReviewTable.js
@@ -172,7 +172,8 @@ export class TaskReviewTable extends Component {
                    "reviewerControls":{permanent: true},
                    "mapperControls":{permanent: true},
                    "viewComments":{},
-                   "tags":{}}
+                   "tags":{},
+                   "additionalReviewers":{}}
 
     let defaultColumns = _keys(columns)
 
@@ -511,6 +512,30 @@ const setupColumnTypes = (props, openComments, data, criteria, pageSize) => {
     ),
   }
 
+  columns.additionalReviewers = {
+    id: 'otherReviewers',
+    Header: props.intl.formatMessage(messages.additionalReviewersLabel),
+    accessor: 'additionalReviewers',
+    sortable: false,
+    filterable: false,
+    maxWidth: 180,
+    Cell: ({row}) => (
+      <div
+        className="row-user-column"
+        style={{color: mapColors(_get(row._original.completedBy, 'username') || row._original.completedBy)}}
+      >
+        {_map(row._original.additionalReviewers, (reviewer, index) => {
+          return (
+            <React.Fragment>
+              <span style={{color: mapColors(reviewer.username)}}>{reviewer.username}</span>
+              {(index + 1) !== _get(row._original.additionalReviewers, 'length') ? ", " : ""}
+            </React.Fragment>
+          )
+        })}
+      </div>
+    ),
+  }
+
   columns.challenge = {
     id: 'challenge',
     Header: makeInvertable(props.intl.formatMessage(messages.challengeLabel),
@@ -777,6 +802,12 @@ const setupColumnTypes = (props, openComments, data, criteria, pageSize) => {
         <Link to={linkTo}>
           {message}
         </Link>
+        {row._original.reviewStatus !== TaskReviewStatus.needed &&
+         row._original.reviewedBy && row._original.reviewedBy.id !== props.user.id &&
+          <div onClick={() => props.history.push(linkTo + "/review", criteria)} className="mr-text-green-lighter hover:mr-text-white mr-cursor-pointer mr-transition">
+            <FormattedMessage {...messages.metaReviewTaskLabel} />
+          </div>
+        }
       </div>
     }
   }