+++ src/services/Search/Search.js
@@ -45,7 +45,7 @@ export const SORT_NAME = 'name'
 export const SORT_CREATED = 'created'
 export const SORT_OLDEST = 'Created'
 export const SORT_POPULARITY = 'popularity'
-export const SORT_COOPERATIVE_WORK = 'has_cooperative_work'
+export const SORT_COOPERATIVE_WORK = 'cooperative_type'
 export const SORT_DEFAULT = 'default'
 export const ALL_SORT_OPTIONS = [SORT_NAME, SORT_CREATED, SORT_OLDEST, SORT_POPULARITY, SORT_COOPERATIVE_WORK, SORT_DEFAULT]